// infrasestructure/interfaces/credits.response.ts
export interface MovieDBCast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string | null;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}

export interface CastResponse {
    id: number;
    cast: MovieDBCast[];
}